Moving algolia serializers to a namespace
Use the algolia serializers when indexing advisers and offices
